requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.14.0
pyyaml>=6.0
pytest>=8.0.0
pytest-cov>=4.1.0
